Remove FlowBuilder/containers/**
