--- src/bp/ui-studio/src/web/views/FlowBuilder/containers/SkillsBuilder.js
@@ -1,19 +1,0 @@
-import { connect } from 'react-redux'
-
-import { cancelNewSkill, insertNewSkill, updateSkill } from '~/actions'
-import SkillsBuilder from '../skills'
-
-const mapStateToProps = state => ({
-  installedSkills: state.skills.installed,
-  ...state.skills.builder
-})
-
-const mapDispatchToProps = {
-  cancelNewSkill,
-  insertNewSkill,
-  updateSkill
-}
-
-const ConnectedSkillsBuilder = connect(mapStateToProps, mapDispatchToProps)(SkillsBuilder)
-
-export default ConnectedSkillsBuilder
